Split limiter test utilities into module
